Initial commit of Mutagen Compose Service.
This is the initial commit of the Mutagen Compose Service. At the
moment, this is just a minimal no-op image that serves as a target for
Mutagen agent executables in a Docker Compose setup. It also only
supports Linux at the moment. As time goes on, additional functionality
and platform support will be added.

Signed-off-by: Jacob Howard <jacob@havoc.io>